requirements.txt
langchain-community>=0.0.21
langchain-openai>=0.0.8
langgraph>=0.0.32
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.0
faiss-cpu>=1.7.4
click>=8.1.0
rich>=13.7.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
